+++ Modules/Intent.Modules.MediatR.DomainEvents/Templates/DomainEventNotification/DomainEventNotificationTemplatePartial.cs
@@ -1,12 +1,12 @@
-using System;
-using System.Collections.Generic;
 using Intent.Engine;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
-using Intent.Modules.DomainEvents.Templates.DomainEventBase;
+using Intent.Modules.Constants;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
+using System;
+using System.Collections.Generic;
 
 [assembly: DefaultIntentManaged(Mode.Merge)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
@@ -34,7 +34,7 @@ namespace Intent.Modules.MediatR.DomainEvents.Templates.DomainEventNotification
 
         private string GetDomainEventBaseType()
         {
-            return GetTypeName(DomainEventBaseTemplate.TemplateId);
+            return GetTypeName(TemplateRoles.Domain.Common.EventBase);
         }
     }
 }