+++ Modules/Intent.Modules.MediatR.DomainEvents/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.2.3
+
+- Improvement: Decoupling this module from the `Intent.Modules.DomainEvents` module so that teams can combine their own Domain Eventing pattners with this one.
+
 ### Version 5.2.2
 
 - Improvement: Updated referenced package version